Move all Fiberplane-specific logic behind a feature flag
