+++ mondrian-charts/Cargo.toml
@@ -10,12 +10,13 @@ rust-version = { workspace = true }
 version = "0.1.0"
 
 [features]
-default = ["image"]
+default = ["fiberplane", "image"]
+fiberplane = ["dep:fiberplane-models"]
 image = ["dep:csscolorparser", "dep:image", "dep:resvg"]
 
 [dependencies]
 csscolorparser = { version = "0.6", optional = true }
-fiberplane-models = { workspace = true }
+fiberplane-models = { workspace = true, optional = true }
 image = { version = "0.24", optional = true }
 resvg = { version = "0.35", optional = true }
 serde = { workspace = true }
