+++ mondrian-charts/src/chart_to_svg/generate_grid_with_axes_svg.rs
@@ -1,8 +1,8 @@
 use super::{ChartOptions, Scales};
-use crate::AbstractChart;
+use crate::MondrianChart;
 
 pub(super) fn generate_grid_with_axes_svg<S, P>(
-    chart: &AbstractChart<S, P>,
+    chart: &MondrianChart<S, P>,
     scales: &Scales,
     options: &ChartOptions,
 ) -> String {
