+++ mondrian-charts/src/fiberplane/generate_stacked_bar_chart_from_timeseries.rs
@@ -1,12 +1,10 @@
 use super::utils::*;
-use crate::types::{
-    AbstractChart, Area, AreaPoint, Axis, Metric, Shape, ShapeList, StackedChartBuckets,
-    Timeseries, TimeseriesSourceData,
-};
+use crate::fiberplane::{Metric, StackedChartBuckets, Timeseries, TimeseriesSourceData};
+use crate::types::{Axis, MondrianChart, Rectangle, Shape, ShapeList};
 
-pub(crate) fn generate_stacked_line_chart_from_timeseries<'source>(
+pub(crate) fn generate_stacked_bar_chart_from_timeseries<'source>(
     input: TimeseriesSourceData<'source, '_>,
-) -> AbstractChart<&'source Timeseries, &'source Metric> {
+) -> MondrianChart<&'source Timeseries, &'source Metric> {
     let BucketsAndAxes {
         mut buckets,
         is_percentage,
@@ -16,10 +14,12 @@ pub(crate) fn generate_stacked_line_chart_from_timeseries<'source>(
 
     let interval = calculate_smallest_time_interval(&buckets);
     if let Some(interval) = interval {
+        x_axis = x_axis.extend_with_interval(interval);
         attach_suggestions_to_x_axis(&mut x_axis, &buckets, interval);
     }
 
-    let mut args = BarArgs {
+    let mut bar_args = BarArgs {
+        bar_width: calculate_bar_width(&x_axis, interval, 1),
         buckets: &mut buckets,
         is_percentage,
         x_axis: &x_axis,
@@ -31,7 +31,11 @@ pub(crate) fn generate_stacked_line_chart_from_timeseries<'source>(
         .iter()
         .map(|timeseries| ShapeList {
             shapes: if timeseries.visible {
-                create_shapes(&timeseries.metrics, interval, &mut args)
+                timeseries
+                    .metrics
+                    .iter()
+                    .filter_map(|metric| create_bar_shape(metric, &mut bar_args))
+                    .collect()
             } else {
                 Vec::new()
             },
@@ -39,7 +43,7 @@ pub(crate) fn generate_stacked_line_chart_from_timeseries<'source>(
         })
         .collect();
 
-    AbstractChart {
+    MondrianChart {
         shape_lists,
         x_axis,
         y_axis,
@@ -47,34 +51,21 @@ pub(crate) fn generate_stacked_line_chart_from_timeseries<'source>(
 }
 
 struct BarArgs<'axes, 'buckets> {
+    bar_width: f64,
     buckets: &'buckets mut StackedChartBuckets,
     is_percentage: bool,
     x_axis: &'axes Axis,
     y_axis: &'axes Axis,
 }
 
-fn create_shapes<'source>(
-    metrics: &'source [Metric],
-    interval: Option<f64>,
-    args: &mut BarArgs,
-) -> Vec<Shape<&'source Metric>> {
-    split_into_continuous_lines(metrics, interval)
-        .iter()
-        .map(|line| {
-            Shape::Area(Area {
-                points: line
-                    .iter()
-                    .filter_map(|metric| create_point_for_metric(metric, args))
-                    .collect(),
-            })
-        })
-        .collect()
-}
-
-fn create_point_for_metric<'source>(
+fn create_bar_shape<'source>(
     metric: &'source Metric,
     args: &mut BarArgs,
-) -> Option<AreaPoint<&'source Metric>> {
+) -> Option<Shape<&'source Metric>> {
+    if metric.value.is_nan() {
+        return None;
+    };
+
     let Some(bucket_value) = args.buckets.get_mut(&metric.time) else {
         return None;
     };
@@ -86,14 +77,17 @@ fn create_point_for_metric<'source>(
         metric.value
     };
 
-    let y_min = bucket_value.current_y;
-    let y_max = y_min + normalize_along_linear_axis(value, args.y_axis);
-    bucket_value.current_y = y_max;
+    let x = normalize_along_linear_axis(time, args.x_axis) - 0.5 * args.bar_width;
+    let y = bucket_value.current_y;
 
-    Some(AreaPoint {
-        x: normalize_along_linear_axis(time, args.x_axis),
-        y_min,
-        y_max,
+    let height = normalize_along_linear_axis(value, args.y_axis);
+    bucket_value.current_y += height;
+
+    Some(Shape::Rectangle(Rectangle {
+        x,
+        y,
+        width: args.bar_width,
+        height,
         source: metric,
-    })
+    }))
 }
