+++ mondrian-charts/src/fiberplane/utils/calculate_bar_x.rs
