+++ mondrian-charts/src/fiberplane/utils/get_y_axis_for_constant_value.rs
