+++ mondrian-charts/src/fiberplane/utils/normalize_along_linear_axis.rs
