+++ mondrian-charts/src/types.rs
@@ -1,87 +1,4 @@
-pub use fiberplane_models::notebooks::{GraphType, StackingType};
-pub use fiberplane_models::providers::{Metric, ProviderEvent, Timeseries};
-pub use fiberplane_models::timestamps::{TimeRange, Timestamp};
 use serde::{Deserialize, Serialize};
-use std::collections::BTreeMap;
-
-/// All the data necessary to generate an abstract chart from an array of
-/// timeseries.
-pub struct TimeseriesSourceData<'source, 'slice> {
-    /// The type of chart to display.
-    pub graph_type: GraphType,
-
-    /// The type of stacking to apply to the chart.
-    pub stacking_type: StackingType,
-
-    /// Array of timeseries data to display in the chart.
-    ///
-    /// Make sure the timeseries contains data for the given time range, or you
-    /// may not see any results.
-    pub timeseries_data: &'slice [&'source Timeseries],
-
-    /// The time range to be displayed.
-    pub time_range: TimeRange,
-}
-
-/// All the data necessary to generate an abstract chart from an array of
-/// timeseries and an array of events.
-///
-/// Note we only support generating line charts from combined timeseries and
-/// events data. If `graph_type` is anything other than [`GraphType::Line`],
-/// the events will be ignored.
-pub struct TimeseriesAndEventsSourceData<'source, 'slice> {
-    /// The type of stacking to apply to the chart.
-    ///
-    /// **Warning:** This property is accepted for consistency, but setting it
-    /// to anything except [GraphType::Line] will cause the events to be
-    /// ignored.
-    pub graph_type: GraphType,
-
-    /// The type of stacking to apply to the chart.
-    pub stacking_type: StackingType,
-
-    /// Array of timeseries data to display in the chart.
-    ///
-    /// Make sure the timeseries contains data for the given time range, or you
-    /// may not see any results.
-    pub timeseries_data: &'slice [&'source Timeseries],
-
-    /// Array of events to display in the chart.
-    pub events: &'slice [&'source ProviderEvent],
-
-    /// The time range to be displayed.
-    pub time_range: TimeRange,
-}
-
-/// Source type for series in charts that contain combined data sources.
-#[derive(Serialize)]
-#[non_exhaustive]
-#[serde(tag = "type", rename_all = "snake_case")]
-pub enum SeriesSource<'source> {
-    Timeseries(&'source Timeseries),
-    Events,
-}
-
-impl<'source> From<&'source Timeseries> for SeriesSource<'source> {
-    fn from(value: &'source Timeseries) -> Self {
-        Self::Timeseries(value)
-    }
-}
-
-/// Source type for points in charts that contain combined data sources.
-#[derive(Serialize)]
-#[non_exhaustive]
-#[serde(tag = "type", rename_all = "snake_case")]
-pub enum PointSource<'source> {
-    Metric(&'source Metric),
-    Event(&'source ProviderEvent),
-}
-
-impl<'source> From<&'source Metric> for PointSource<'source> {
-    fn from(value: &'source Metric) -> Self {
-        Self::Metric(value)
-    }
-}
 
 /// An abstract chart with information about what to render and where to render
 /// it.
@@ -96,22 +13,10 @@ impl<'source> From<&'source Metric> for PointSource<'source> {
 /// [Timeseries] and [Metric], respectively.
 #[derive(Default, Deserialize, Serialize)]
 #[serde(rename_all = "camelCase")]
-pub struct AbstractChart<S, P> {
+pub struct MondrianChart<S, P> {
     pub x_axis: Axis,
     pub y_axis: Axis,
     pub shape_lists: Vec<ShapeList<S, P>>,
-}
-
-impl<'source> From<AbstractChart<&'source Timeseries, &'source Metric>>
-    for AbstractChart<SeriesSource<'source>, PointSource<'source>>
-{
-    fn from(value: AbstractChart<&'source Timeseries, &'source Metric>) -> Self {
-        Self {
-            shape_lists: value.shape_lists.into_iter().map(ShapeList::into).collect(),
-            x_axis: value.x_axis,
-            y_axis: value.y_axis,
-        }
-    }
 }
 
 /// Defines the range of values that are displayed along a given axis.
@@ -146,45 +51,6 @@ impl Axis {
     }
 }
 
-/// A map that holds arbitrary values per timestamp.
-///
-/// The keys in the map represent timestamps, while the values are generic and
-/// are calculated over all the metrics matching that timestamp.
-pub(crate) type Buckets<T> = BTreeMap<Timestamp, T>;
-
-/// Represents the minimum and maximum values inside a series of numbers.
-#[derive(Clone, Copy)]
-pub(crate) struct MinMax(pub f64, pub f64);
-
-impl MinMax {
-    pub fn from_value(value: f64) -> Self {
-        Self(value, value)
-    }
-
-    #[must_use]
-    pub fn extend_with_other(mut self, other: Self) -> Self {
-        if other.0 < self.0 {
-            self.0 = other.0;
-        }
-        if other.1 > self.1 {
-            self.1 = other.1;
-        }
-
-        self
-    }
-
-    #[must_use]
-    pub fn extend_with_value(mut self, value: f64) -> Self {
-        if value < self.0 {
-            self.0 = value;
-        } else if value > self.1 {
-            self.1 = value;
-        }
-
-        self
-    }
-}
-
 /// List of shapes that belongs together.
 ///
 /// These should be rendered in the same color.
@@ -200,17 +66,6 @@ pub struct ShapeList<S, P> {
     pub source: S,
 }
 
-impl<'source> From<ShapeList<&'source Timeseries, &'source Metric>>
-    for ShapeList<SeriesSource<'source>, PointSource<'source>>
-{
-    fn from(value: ShapeList<&'source Timeseries, &'source Metric>) -> Self {
-        Self {
-            shapes: value.shapes.into_iter().map(Shape::into).collect(),
-            source: value.source.into(),
-        }
-    }
-}
-
 #[derive(Deserialize, Serialize)]
 #[serde(tag = "type", rename_all = "snake_case")]
 pub enum Shape<P> {
@@ -220,17 +75,6 @@ pub enum Shape<P> {
     Rectangle(Rectangle<P>),
 }
 
-impl<'source> From<Shape<&'source Metric>> for Shape<PointSource<'source>> {
-    fn from(other: Shape<&'source Metric>) -> Self {
-        match other {
-            Shape::Area(area) => Shape::Area(area.into()),
-            Shape::Line(line) => Shape::Line(line.into()),
-            Shape::Point(point) => Shape::Point(point.into()),
-            Shape::Rectangle(rectangle) => Shape::Rectangle(rectangle.into()),
-        }
-    }
-}
-
 /// An area to be drawn between two lines that share their X coordinates.
 ///
 /// Area points move from left to right.
@@ -238,14 +82,6 @@ impl<'source> From<Shape<&'source Metric>> for Shape<PointSource<'source>> {
 #[serde(rename_all = "camelCase")]
 pub struct Area<P> {
     pub points: Vec<AreaPoint<P>>,
-}
-
-impl<'source> From<Area<&'source Metric>> for Area<PointSource<'source>> {
-    fn from(value: Area<&'source Metric>) -> Self {
-        Self {
-            points: value.points.into_iter().map(AreaPoint::into).collect(),
-        }
-    }
 }
 
 /// A single data point in an area shape.
@@ -267,32 +103,12 @@ pub struct AreaPoint<P> {
     pub source: P,
 }
 
-impl<'source> From<AreaPoint<&'source Metric>> for AreaPoint<PointSource<'source>> {
-    fn from(value: AreaPoint<&'source Metric>) -> Self {
-        Self {
-            x: value.x,
-            y_min: value.y_min,
-            y_max: value.y_max,
-            source: value.source.into(),
-        }
-    }
-}
-
 /// A line to be drawn between two or more points.
 #[derive(Deserialize, Serialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Line<P> {
     pub points: Vec<Point<P>>,
 }
-
-impl<'source> From<Line<&'source Metric>> for Line<PointSource<'source>> {
-    fn from(value: Line<&'source Metric>) -> Self {
-        Self {
-            points: value.points.into_iter().map(Point::into).collect(),
-        }
-    }
-}
-
 /// A single point in the chart.
 ///
 /// Points can be rendered independently as a dot, or can be used to draw lines
@@ -312,16 +128,6 @@ pub struct Point<P> {
     pub source: P,
 }
 
-impl<'source> From<Point<&'source Metric>> for Point<PointSource<'source>> {
-    fn from(value: Point<&'source Metric>) -> Self {
-        Self {
-            x: value.x,
-            y: value.y,
-            source: value.source.into(),
-        }
-    }
-}
-
 /// A rectangle to be rendered inside the chart.
 #[derive(Deserialize, Serialize)]
 #[serde(rename_all = "camelCase")]
@@ -335,28 +141,4 @@ pub struct Rectangle<P> {
     ///
     /// This would be a [Metric] if the chart was generated from [Timeseries].
     pub source: P,
-}
-
-impl<'source> From<Rectangle<&'source Metric>> for Rectangle<PointSource<'source>> {
-    fn from(value: Rectangle<&'source Metric>) -> Self {
-        Self {
-            x: value.x,
-            y: value.y,
-            width: value.width,
-            height: value.height,
-            source: value.source.into(),
-        }
-    }
-}
-
-pub(crate) type StackedChartBuckets = Buckets<StackedChartBucketValue>;
-
-#[derive(Clone, Copy, Default)]
-pub(crate) struct StackedChartBucketValue {
-    /// Used to keep track of how much a bucket is "filled" while calculating
-    /// the area shapes.
-    pub current_y: f64,
-
-    /// The sum of all values in the bucket.
-    pub total: f64,
 }
